+++ src/components/organisms/Header/TagWallPage/Header.scss
@@ -1,0 +1,6 @@
+.Header {
+    width: 100%;
+    height: 134px;
+    flex-shrink: 0;
+    background-color: rgba(0, 163, 255, 0.31);
+}
