+++ src/components/organisms/Header/TagWallPage/Header.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import "./Header.scss";
+
+export default class Header extends React.Component {
+  render() {
+    return <header className="Header"></header>;
+  }
+}
