+++ src/components/organisms/Main/IndexPage/Main.scss
