+++ src/components/pages/IndexPage.tsx
@@ -1,7 +1,10 @@
 import React from "react";
+import Main from "../organisms/Main/IndexPage/Main";
+import Header from "../organisms/Header/IndexPage/Header";
 import MediaQuery from "react-responsive";
+import Footer from "../organisms/Footer/IndexPage/Footer";
 
-export default class TagWall extends React.Component {
+export default class Index extends React.Component {
   render() {
     return (
       <div className="Index">
@@ -11,8 +14,12 @@ export default class TagWall extends React.Component {
         </MediaQuery>
         <MediaQuery query="(min-width: 768px)">
           {/* // iPadの横幅を分岐点にしている */}
-          <div className="App-allwrapper"></div>
-          <div>ここにPC向けの要素を書く</div>
+          <div className="App-allwrapper">
+            <Header />
+            <Main />
+            <Footer />
+          </div>
+          {/* <div>ここにPC向けの要素を書く</div> */}
         </MediaQuery>
       </div>
     );
